akt: autobid cpu params supports float values, add app level auth

+++ client/js/akt/AKTDashboard.js
@@ -547,7 +547,7 @@ export class AKTDashboard {
 		*/
 		$('#runSave').off('click').on('click',()=>{
 			const fees = $('#runFees').val() == '' ? 10000 : parseInt($('#runFees').val());
-			const cpu = $('#cpuPrice').val() == '' ? 10 : parseInt($('#cpuPrice').val());
+			const cpu = $('#cpuPrice').val() == '' ? 10 : parseFloat($('#cpuPrice').val());
 			const autostart = $('#autostart').is(':checked');
 			$('#runSave .foreground').html('Starting Up...');
 			$('#runSave .background').html('Starting Up...');
